requirements.txt
flask
opencv-python
mediapipe
protobuf
numpy
pandas
Pillow
tensorflow
